Update syntax for setting suggested sort with praw 5
See comment on https://github.com/mattabullock/Baseball-GDT-Bot/pull/35.
This is necessary because there are two pull requests pending: one to
set suggested sort for pre and post threads in the current praw 3
syntax, and another to upgrade to praw 5 syntax. Whichever is merged
second will need this update made.